+++ src/main/java/pyre/coloredredstone/blocks/BlockColoredRedstoneLamp.java
@@ -186,7 +186,20 @@ public class BlockColoredRedstoneLamp extends Block implements IColoredFeatures,
     @Override
     public void onEntityWalk(World worldIn, BlockPos pos, Entity entityIn) {
         super.onEntityWalk(worldIn, pos, entityIn);
-        withering(worldIn, entityIn, getColor(worldIn, pos));
+
+        if (!worldIn.isRemote && entityIn instanceof EntityLivingBase && (worldIn.getWorldTime() % 20 == 0)) {
+            EnumColor color = getColor(worldIn, pos);
+
+            if (color.equals(WITHERING_COLOR)) {
+                withering(worldIn, entityIn);
+            } else if (color.equals(SLUGGISH_COLOR)) {
+                sluggish(worldIn, entityIn);
+            } else if (color.equals(SPEEDY_COLOR)) {
+                speedy(worldIn, entityIn);
+            } else if (color.equals(HEALTHY_COLOR)) {
+                healthy(worldIn, entityIn);
+            }
+        }
     }
 
     @SideOnly(Side.CLIENT)